Make fbdev device node runtime configurable via environment variable
Test whether $LV_LINUX_FBDEV_DEVICE environment variable is non-NULL and
in case it is, use it as the video card file in lv_linux_fbdev_set_file().
Otherwise fall back to "/dev/fb0", i.e. the current behavior. This way,
it is possible to test LVGL on systems with multiple fbdev devices.

Signed-off-by: Marek Vasut <marex@denx.de>